Basic annotations - floor

+++ 0x00-python_variable_annotations/2-floor.py
@@ -5,7 +5,7 @@ a type-annotated function floor which takes a float n
 as argument and returns the floor of the float.
 
 Functions:
-    floor(n: float) -> float
+    floor(n: float) -> int
 """
